requirements.txt
bcrypt==4.2.0
frontend
langchain
langchain-community
langchain-core
langchain-openai
langchain==0.3.1
langchain_core
langchain-mongodb
numpy
numpy==1.26.4
openai
open-clip-torch
chromadb==0.6.2
comtypes
opencv-python-headless
pillow==10.4.0
prettyprint
pdfplumber
pymongo==4.10.1
PyMuPDF
PyMuPDF
pytesseract==0.3.13
python-docx
scikit-learn
python-dotenv==1.0.1
streamlit==1.39.0
streamlit_folium==0.23.0
tiktoken
transformers==4.45.0
pdf2image
torch